requirements.txt

pandas
numpy
scikit-learn
umap-learn
plotly
kagglehub
